refactor: core:domain의 모델 클래스들을 core:model 모듈로 이동
- `core/domain`에 있던 모델 파일들을 `core/model` 모듈의 패키지 구조에 맞춰 이동 및 패키지명 변경
- `UserRepository`, `GetReceiversUseCase` 등 관련 클래스의 import 문 업데이트
- `core/domain` 및 `core/model`의 `build.gradle.kts` 의존성 정리
