+++ core/model/build.gradle.kts
@@ -9,6 +9,4 @@ android {
 }
 
 dependencies {
-    implementation(projects.core.model)
-    implementation(projects.core.common)
 }
