+++ core/model/src/main/kotlin/com/afternote/core/model/DailyQuestionModel.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.domain.model
+package com.afternote.core.model
 
 data class DailyQuestionAnswerItem(
     val dailyQuestionAnswerId: Long,
