+++ core/model/src/main/kotlin/com/afternote/core/model/MindRecordModel.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.domain.model
+package com.afternote.core.model
 
 data class ReceiverMindRecordItem(
     val recordId: Long,
