+++ core/model/src/main/kotlin/com/afternote/core/model/UserProfileModel.kt
@@ -1,4 +1,4 @@
-package com.afternote.core.domain.model
+package com.afternote.core.model
 
 data class UserProfileModel(
     val name: String,
